aislesense_navigator/requirements.txt
# AisleSense Region Navigator — Python dependencies
# Install:  pip install -r requirements.txt

Pillow>=9.0
PyYAML>=6.0
numpy>=1.20
opencv-python>=4.8
# Headless alternative (no GUI): use opencv-python-headless>=4.8 instead

# rclpy is NOT pip‑installable — it comes with your ROS 2 installation.
# Source ROS 2 before running if you want live navigation:
#   source /opt/ros/humble/setup.bash
# The app works without rclpy (map editing / region saving still works,
# navigation falls back to `ros2 topic pub` via subprocess).
